requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg[binary]
alembic
python-jose
passlib[bcrypt]
python-dotenv
passlib[bcrypt]==1.7.4
bcrypt==3.2.0
python-multipart